+++ src/main/java/com/example/demo/controller/ClientController.java
@@ -23,11 +23,12 @@ import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.example.demo.model.Client;
+import com.example.demo.model.PharmFavorite;
 import com.example.demo.model.Survey;
-import com.example.demo.model.SurveyFeedback;
 import com.example.demo.model.SurveyFeedbackJoin;
 import com.example.demo.service.ChatGptService;
 import com.example.demo.service.ClientService;
+import com.example.demo.service.PharmFavoriteService;
 import com.example.demo.service.SurveyFeedbackService;
 import com.example.demo.service.SurveyService;
 import com.example.demo.service.VerifyService;
@@ -53,14 +54,6 @@ public class ClientController {
 	@Autowired
 	private VerifyService verifyService;
 	
-	@Autowired
-	private SurveyService surveyService;
-	
-	@Autowired
-	private ChatGptService chatGptService;
-	
-	@Autowired
-	private SurveyFeedbackService surveyFeedbackService;
 
 	// 회원가입 화면
 	@GetMapping("/register")
@@ -188,177 +181,12 @@ public class ClientController {
 		return "redirect:/login";
 	}
 
-//******************--------------------------------    
-
-// 마이페이지 추가
-
-	@GetMapping("/mypage")
-	public String mypage(HttpSession session, Model model) {
-	    Client client = (Client) session.getAttribute("loginClient");
-	    model.addAttribute("client", client);
-
-	    // 관심사 유효성 체크
-	    if (client == null || client.getInterest() == null || client.getInterest().isEmpty()) {
-	        model.addAttribute("recommendExplanation", "관심사가 등록되지 않았습니다.");
-	        return "client/mypage";
-	    }
-
-	    String interestStr = client.getInterest();
-	    String[] interestArr = interestStr.split(",");
-
-	    // 설문 추천
-	    List<Survey> recommendedSurveys = surveyService.getRecommendedSurveys(interestArr);
-	    model.addAttribute("recommendedSurveys", recommendedSurveys);
-
-	    // GPT 설명 캐시 로직
-	    if (interestArr.length > 0 && !recommendedSurveys.isEmpty()) {
-	        try {
-	            String explanation = chatGptService.getOrCreateSurveyExplanation(
-	                client.getClientId(), interestArr, recommendedSurveys
-	            );
-	            model.addAttribute("recommendExplanation", explanation);
-	        } catch (Exception e) {
-	            e.printStackTrace();
-	            model.addAttribute("recommendExplanation", "GPT 설명 생성 실패");
-	            System.out.println("GPT 오류 응답 내용: " + e.getMessage());
-	        }
-	    }
-
-	    // 가입일로부터 D+ 계산
-	    if (client.getRegDate() != null) {
-	        LocalDate joinDate = client.getRegDate().toLocalDateTime().toLocalDate();
-	        LocalDate today = LocalDate.now(ZoneId.systemDefault());
-	        long dday = ChronoUnit.DAYS.between(joinDate, today);
-	        model.addAttribute("dDay", dday);
-	    }
-
-	    return "client/mypage";
-	}
-
-	// 마이페이지 프로필 화면
-	@GetMapping("/mypageProfile")
-	public String mypageProfile(HttpSession session, Model model) {
-		Client client = (Client) session.getAttribute("loginClient");
-		model.addAttribute("client", client);
-		return "client/mypageProfile"; // /WEB-INF/views/client/mypageProfile.jsp
-	}
 	
-	// 설문조사 결과
-	@GetMapping("/mypageReport")
-	public String mypageReport(HttpSession session, Model model) {
-	    Client client = (Client) session.getAttribute("loginClient");
-
-	    if (client == null) {
-	        return "redirect:/login"; // 비로그인 사용자는 로그인 페이지로
-	    }
-
-	    String clientId = client.getClientId();
-	    List<SurveyFeedbackJoin> feedbackList = surveyFeedbackService.getFeedbackByClientId(clientId);
-	    model.addAttribute("client", client);
-	    model.addAttribute("feedbackList", feedbackList);
-	    return "client/mypageReport";
-	}
-
-
-	// 마이페이지 회원 정보 수정 화면
-	@GetMapping("/mypageUpdate")
-	public String mypageUpdate(HttpSession session, Model model) {
-	    Client sessionClient = (Client) session.getAttribute("loginClient");
-	    if (sessionClient == null) {
-	        return "redirect:/login"; 
-	    }
-	    Client client = clientService.getClientById(sessionClient.getClientId());
-
-	    String interestStr = client.getInterest();
-	    String[] interestArr = interestStr != null ? interestStr.split(",") : new String[0];
-	    System.out.println("mypageUpdate - client: " + client.toString());
-	    model.addAttribute("interestList", Arrays.asList(interestArr));
-	    model.addAttribute("client", client);
-
-	    return "client/mypageUpdate";
-	}
-
-	// 마이페이지 비밀번호 변경 화면
-	@GetMapping("/mypagePassword")
-	public String mypagePassword(HttpSession session, Model model) {
-		Client client = (Client) session.getAttribute("loginClient");
-		model.addAttribute("loginClient", client);
-		return "client/mypagePassword";
-	}
-
-	// 마이페이지 회원 탈퇴 화면
-	@GetMapping("/mypageDelete")
-	public String mypageDelete(HttpSession session, Model model) {
-		Client client = (Client) session.getAttribute("loginClient");
-		model.addAttribute("loginClient", client);
-		return "client/mypageDelete";
-	}
+	
+	
 
 	// 회원 수정 폼
-	@GetMapping("/update")
-	public String showUpdateForm(HttpSession session, Model model) {
-		Client loginClient = (Client) session.getAttribute("loginClient");
-
-		if (loginClient == null) {
-			return "redirect:/client/login"; // 로그인 안 했으면 로그인으로
-		}
-
-		model.addAttribute("client", loginClient); // JSP에 회원 정보 넘겨줌
-		return "client/update"; // 수정 폼 화면 반환
-	}
-
-	// 회원 정보 수정 처리
-	@PostMapping("/update")
-	public String updateClient(@ModelAttribute Client client, 
-			  				   @RequestParam(value = "interestList", required = false) List<String> interestList,
-							   HttpSession session,
-			                   RedirectAttributes redirectAttributes) {
-		
-		Client loginClient = (Client) session.getAttribute("loginClient");
-
-		if (loginClient == null) {
-			return "redirect:/client/login"; // 로그인 안 한 경우 login폼으로 이동
-		}
-
-		// 세션의 ID로 고정 (보안상 중요!)
-		client.setClientId(loginClient.getClientId());
-		
-
-	    // ✅ 일반 로그인 사용자는 성별 수정 불가 (기존 값 유지)
-	    if (!"N".equals(loginClient.getGender())) {
-	        client.setGender(loginClient.getGender());
-	    }
-
-	    // ✅ 관심사 체크박스를 문자열로 병합
-	    if (interestList != null && !interestList.isEmpty()) {
-	        client.setInterest(String.join(",", interestList));
-	    } else {
-	        client.setInterest(""); // 아무것도 선택 안 한 경우
-	    }
-
-
-		boolean result = clientService.updateClient(client);
-		
-		if (result) {
-			 // ✅ DB에서 최신 회원 정보 가져옴
-		    Client refreshed = clientService.getClientById(client.getClientId());
-
-		    // ✅ 로그로 관심사 확인
-		    System.out.println("[DEBUG] 수정 후 DB에서 가져온 관심사: " + refreshed.getInterest());
-
-		    // ✅ 세션에 최신 정보 저장
-		    session.setAttribute("loginClient", refreshed);
-
-		    redirectAttributes.addFlashAttribute("message", "회원 정보가 성공적으로 수정되었습니다.");
-		    return "redirect:/client/mypageProfile";
-		} else {
-			// 실패 메시지 추가
-			redirectAttributes.addFlashAttribute("error", "회원 정보 수정에 실패했습니다.");
-
-			// 수정 페이지로 리다이렉트
-			return "redirect:/client/mypageUpdate";
-		}
-	}
+	
 
 	// 비밀번호 변경 폼
 	@GetMapping("/changePassword")
@@ -371,7 +199,7 @@ public class ClientController {
 		}
 
 		model.addAttribute("loginClient", client);
-		return "client/mypagePassword";
+		return "mypage/mypagePassword";
 	}
 
 	// 비밀번호 변경 처리
@@ -385,20 +213,20 @@ public class ClientController {
 		// 현재 비밀번호 확인
 		if (!passwordEncoder.matches(currentPassword, client.getPassword())) {
 			redirectAttributes.addFlashAttribute("error", "❌ 현재 비밀번호가 일치하지 않습니다.");
-			return "redirect:/client/mypagePassword";
+			return "redirect:/mypage/mypagePassword";
 		}
 
 		// 새 비밀번호 유효성 검사
 		String pwRegex = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[!@#$%^&*()\\-_=+\\[\\]{};:'\",.<>/?`~]).{8,}$";
 		if (!newPassword.matches(pwRegex)) {
 			redirectAttributes.addFlashAttribute("error", "❌ 비밀번호는 8자 이상, 영문+숫자+특수문자를 포함해야 합니다.");
-			return "redirect:/client/changePassword";
+			return "redirect:/mypage/changePassword";
 		}
 
 		// 새 비밀번호 일치 여부 확인
 		if (!newPassword.equals(confirmPassword)) {
 			redirectAttributes.addFlashAttribute("error", "❌ 새 비밀번호가 서로 다릅니다.");
-			return "redirect:/client/mypagePassword";
+			return "redirect:/mypage/mypagePassword";
 		}
 
 		// 비밀번호 업데이트
@@ -406,7 +234,7 @@ public class ClientController {
 		clientService.updatePassword(client);
 
 		redirectAttributes.addFlashAttribute("success", " 비밀번호가 성공적으로 변경되었습니다. ");
-		return "redirect:/client/mypage";
+		return "redirect:/mypage/";
 	}
 
 	// ✅ 현재 비밀번호 일치 확인용 (AJAX 요청 처리)
