+++ src/main/java/com/example/demo/model/PharmFavorite.java
@@ -1,0 +1,15 @@
+package com.example.demo.model;
+
+import lombok.Data;
+
+@Data
+public class PharmFavorite {
+	
+	private String clientId;
+	private String pharmId;
+	private String pharmName;
+	private String pharmAddress;
+	private String pharmTel;
+	private String openHour;
+	
+}
