+++ src/main/webapp/WEB-INF/views/includes/header.jsp
@@ -17,7 +17,7 @@
 			<c:choose>
 				<c:when test="${not empty sessionScope.loginClient}">
 					<a href="/client/logout">로그아웃</a>
-					<a href="/client/mypage">마이페이지</a>
+					<a href="/mypage/">마이페이지</a>
 				</c:when>
 				<c:when test="${not empty sessionScope.loginCounselor}">
 					<a href="/client/logout">로그아웃</a>
