+++ src/main/webapp/WEB-INF/views/mypage/mypageProfile.jsp
@@ -1,40 +1,37 @@
 <%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <!DOCTYPE html>
 <html>
 <head>
-<title>회원탈퇴</title>
-<link rel="stylesheet"
-	href="${pageContext.request.contextPath}/css/mypage.css">
+<meta charset="UTF-8">
+<title>내 프로필</title>
 <link rel="stylesheet" href="${pageContext.request.contextPath}/css/client.css">
+<link rel="stylesheet" href="${pageContext.request.contextPath}/css/mypage.css">
 <link rel="stylesheet" href="${pageContext.request.contextPath}/css/common.css">
+
 </head>
 <body>
 <!-- 헤더 영역 -->
 <jsp:include page="/WEB-INF/views/includes/header.jsp" flush="true"/>
-
-	<!-- 회원탈퇴 페이지의 맨 위나 body 안에 삽입 -->
-	<c:if test="${not empty error}">
+	<c:if test="${not empty message}">
 		<script>
-			alert("${error}");
+			alert("${message}");
 		</script>
 	</c:if>
 
 	<!-- ✅ 마이페이지 wrapper 안에 타이틀 포함 -->
 	<div class="mypage-wrapper">
 
-		<!-- ✅ 사이드바 -->
+		<!-- ✅사이드바 -->
 		<%
 		String uri = request.getRequestURI();
 		%>
 		<div class="mypage-sidebar">
 			<h1 class="mypage-title">
-				<a href="${pageContext.request.contextPath}/mypage/mypage"
+				<a href="${pageContext.request.contextPath}/mypage/"
 					style="text-decoration: none; color: inherit;">마이페이지</a>
 			</h1>
 
-			<div class="mypage-divider"></div>
-			
 			<div class="menu-item" onclick="location.href='${pageContext.request.contextPath}/mypage/mypageReservation'">예약확인</div>
 			<div class="menu-item" onclick="location.href='${pageContext.request.contextPath}/mypage/mypageReport'">검사기록</div>
 			<div class="menu-item" onclick="location.href='${pageContext.request.contextPath}/mypage/mypagePharmFav'">즐겨찾기</div>
@@ -44,49 +41,46 @@
 			<div class="menu-item" onclick="location.href='${pageContext.request.contextPath}/mypage/mypageDelete'">회원탈퇴</div>
 		</div>
 
-
 		<!-- ✅ 콘텐츠 영역 -->
 		<div class="mypage-content">
-			<h2 class="title">회원 탈퇴</h2>
+
+			<h2 class="title">내 프로필</h2>
 			<div class="divider"></div>
 
-			<form action="/mypage/delete" method="post" onsubmit="return validateDeleteForm();">
+			<!-- 아이디 -->
+			<div class="form-group">
+				<label>아이디</label> <input type="text" value="${client.clientId}" readonly>
+			</div>
 
-				<div class="form-group">
-					<label>ID</label> <input type="text" name="id"
-						value="${loginClient.clientId}" readonly>
-				</div>
+			<!-- 이름 -->
+			<div class="form-group">
+				<label>이름</label> <input type="text" value="${client.name}" readonly>
+			</div>
 
-				<div class="form-group">
-					<label>비밀번호</label> 
-					<input type="password" name="pw" id="pw" required>
-				</div>
 
-				<div class="form-group">
-					<label>비밀번호 확인</label> 
-					<input type="password" name="pwConfirm" id="pwConfirm" oninput="checkPasswordMatch()">
+			<!-- 전화번호 -->
+			<div class="form-group">
+				<label>휴대폰번호</label> <input type="text" value="${client.phone}"
+					readonly>
+			</div>
 
-				</div>
-				<!-- 자바스크립트 결과 메시지 표시 -->
-				<p id="pwMatchResult" class="result-text"></p>
+			<!-- 이메일 -->
+			<div class="form-group">
+				<label>이메일</label> <input type="text"
+					value="${client.emailId}@${client.emailDomain}" readonly>
+			</div>
 
-				<div class="form-group button-row">
-					<button type="submit" class="btn btn-half">탈퇴하기</button>
-					<button type="button" class="btn btn-half"
-						onclick="location.href='/mypage/mypage'">취소</button>
-				</div>
+			<!-- 생년월일 -->
+			<div class="form-group">
+				<label>생년월일</label> <input type="text" value="${client.birth}"
+					readonly>
+			</div>
 
-			</form>
 		</div>
-		
-		</div>
 
-		<div style="height:300px"></div>
-		<!-- js연결! -->
-		<script src="https://code.jquery.com/jquery-latest.js"></script>
-		<script src="${pageContext.request.contextPath}/js/mypageDelete.js"></script>
-		
-		<!-- 프로필 아이콘 드롭박스 -->
+	</div>
+	<!-- 프로필 아이콘 드롭박스 -->
+	<script src="<c:url value='/js/proFile.js'/>"></script>
 <!-- 푸터 영역 -->
 <jsp:include page="/WEB-INF/views/includes/footer.jsp" flush="true"/>
 </body>
