+++ src/main/webapp/WEB-INF/views/survey/clientResult.jsp
@@ -19,7 +19,7 @@
     </div>
 
     <div class="history-link">
-      <a href="${pageContext.request.contextPath}/client/mypageReport">검사 기록 보기</a>
+      <a href="${pageContext.request.contextPath}/mypage/mypageReport">검사 기록 보기</a>
     </div>
   </div>
 </body>
